requirements-test.txt
pytest-homeassistant-custom-component
pytest-asyncio
numpy
pandas
astral
